+++ sync-for-payables/src/sdk/bills.ts
@@ -8,6 +8,7 @@ import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
+import jp from "jsonpath";
 
 /**
  * Bills
@@ -873,6 +874,20 @@ export class Bills {
                 break;
         }
 
+        res.next = async (): Promise<operations.ListBillsResponse | null> => {
+            const nextCursor = jp.value(JSON.parse(decodedRes), "");
+            if (nextCursor === undefined) {
+                return null;
+            }
+
+            return await this.list(
+                {
+                    ...req,
+                },
+                retries,
+                config
+            );
+        };
         return res;
     }